+++ NEWS.md
@@ -1,5 +1,6 @@
 # meteoR 0.0.9101
 - New function: GetTopography()
+- ReadNetCDF() now can output arrays and vectors. 
 
 # meteoR 0.0.9100
 - Full (not perfect) documentation.